Move groovy map support to separate class
