+++ plugin/jvm/src/main/java/com/twosigma/beaker/easyform/ObservableMap.java
@@ -1,0 +1,69 @@
+package com.twosigma.beaker.easyform;
+
+import java.util.Collection;
+import java.util.Map;
+import java.util.Observable;
+import java.util.Set;
+
+public abstract class ObservableMap<K, V> extends Observable implements Map<K, V> {
+
+  @Override
+  public int size() {
+    return 0;
+  }
+
+  @Override
+  public boolean isEmpty() {
+    return false;
+  }
+
+  @Override
+  public boolean containsKey(Object key) {
+    return false;
+  }
+
+  @Override
+  public boolean containsValue(Object value) {
+    return false;
+  }
+
+  @Override
+  public V get(Object key) {
+    return null;
+  }
+
+  @Override
+  public V put(K key, V value) {
+    return null;
+  }
+
+  @Override
+  public V remove(Object key) {
+    return null;
+  }
+
+  @Override
+  public void putAll(Map<? extends K, ? extends V> m) {
+
+  }
+
+  @Override
+  public void clear() {
+
+  }
+
+  @Override
+  public Set<K> keySet() {
+    return null;
+  }
+
+  @Override
+  public Collection<V> values() {
+    return null;
+  }
+
+  @Override
+  public Set<Entry<K, V>> entrySet() {
+    return null;
+  }
+}
